+++ src/TestHelpers/Properties/AssemblyInfo.cs
@@ -4,7 +4,7 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Test")]
+[assembly: AssemblyTitle("TestHelpers")]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("1798b1e6-6e75-4ed5-bcbd-3f4bb43deb08")]
